requirements.txt
feedparser
newspaper3k
transformers
torch
spacy
gTTS

fastapi
uvicorn

streamlit
requests
gtts
